ABFcore: Readout the FileGUID for ABF2 properly
Comparing with ABFInfo has shown that the FileGUID is not just a number
as currently readout but a hex string with 128bit.

With some trial and error and the hint, from stimfit, that ABFv1 uses
CoCreateGUID [1] it was possible to reverse engineer the exact
formatting of the 128bit GUID data.

[1]: https://docs.microsoft.com/en-us/windows/desktop/api/combaseapi/nf-combaseapi-cocreateguid

Unfortunately at least for the files accessible to the author the GUID
is not unique for all files.